+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Removed
+ - Drop Python 3.5 support
 
 ## [1.8.0] - 2022-02-26
 ### Added